# api/openapi.yaml
openapi: 3.0.3
info:
  title: 接机管理系统 API
  description: 微信小程序接机管理系统的后端API文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@pickup.com
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api.pickup.com/v1
    description: 生产环境
paths:
...
  /auth/wechat/login:
    post:
      tags:
        - 认证
      summary: 微信登录
      description: 通过微信小程序code和手机号code进行登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - phone_code
              properties:
                code:
                  type: string
                  description: 微信登录code
                  example: "081234567890"
                phone_code:
                  type: string
                  description: 手机号授权code
                  example: "081234567890"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 0
                message: "success"
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    id: 1
                    openid: "o1234567890"
                    phone: "13800138000"
                    nickname: "微信用户"
                    role: "passenger"
                    status: "active"
                time: "2024-01-01T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/me:
    get:
      tags:
        - 认证
      summary: 获取当前用户信息
      description: 获取当前登录用户的详细信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # 报名相关
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required:
        - code
        - message
        - time
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: "success"
        data:
          type: object
          description: 响应数据
        time:
          type: string
          format: date-time
          description: 响应时间（ISO8601格式）
          example: "2024-01-01T12:00:00Z"
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 10001
            message: "invalid parameters"
            time: "2024-01-01T12:00:00Z"
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 10002
            message: "unauthorized"
            time: "2024-01-01T12:00:00Z"
...
    InternalError:
      description: 内部服务器错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 10006
            message: "internal server error"
            time: "2024-01-01T12:00:00Z"
...